+++ transport/x/grpc/grpcheader/grpcheader.go
@@ -25,12 +25,7 @@
 // as well as for future cross-package yarpc work.
 package grpcheader
 
-import (
-	"context"
-	"strings"
-
-	"google.golang.org/grpc/metadata"
-)
+import "strings"
 
 // these are the same as in transport/http but lowercase
 // http2 does all lowercase headers and this should be explicit
@@ -78,59 +73,4 @@ var (
 func IsReserved(header string) bool {
 	_, ok := reservedHeaders[strings.ToLower(header)]
 	return ok
-}
-
-// ContextWrapper wraps a context for grpc-go with the required headers for yarpc.
-//
-// This is a convenience object for use when using grpc-go clients.
-type ContextWrapper struct {
-	md metadata.MD
-}
-
-// NewContextWrapper returns a new ContextWrapper.
-//
-// The only fields that a grpc-go client needs to set are caller and service.
-func NewContextWrapper() *ContextWrapper {
-	return &ContextWrapper{metadata.New(nil)}
-}
-
-// Wrap wraps the given context with the headers.
-func (c *ContextWrapper) Wrap(ctx context.Context) context.Context {
-	return metadata.NewOutgoingContext(ctx, c.md)
-}
-
-// WithCaller returns a new ContextWrapper with the given caller.
-func (c *ContextWrapper) WithCaller(caller string) *ContextWrapper {
-	return c.copyAndAdd(CallerHeader, caller)
-}
-
-// WithService returns a new ContextWrapper with the given service.
-func (c *ContextWrapper) WithService(service string) *ContextWrapper {
-	return c.copyAndAdd(ServiceHeader, service)
-}
-
-// WithShardKey returns a new ContextWrapper with the given shard key.
-func (c *ContextWrapper) WithShardKey(shardKey string) *ContextWrapper {
-	return c.copyAndAdd(ShardKeyHeader, shardKey)
-}
-
-// WithRoutingKey returns a new ContextWrapper with the given routing key.
-func (c *ContextWrapper) WithRoutingKey(routingKey string) *ContextWrapper {
-	return c.copyAndAdd(RoutingKeyHeader, routingKey)
-}
-
-// WithRoutingDelegate returns a new ContextWrapper with the given routing delegate.
-func (c *ContextWrapper) WithRoutingDelegate(routingDelegate string) *ContextWrapper {
-	return c.copyAndAdd(RoutingDelegateHeader, routingDelegate)
-}
-
-func (c *ContextWrapper) copyAndAdd(key string, value string) *ContextWrapper {
-	md := c.md
-	if md == nil {
-		md = metadata.New(nil)
-	} else {
-		md = md.Copy()
-	}
-	md[key] = []string{value}
-	return &ContextWrapper{md}
 }
